support for json hydration format (Pages Router) and raw rsc
